requirements.txt
# HeritageArt CNN-ViT Hybrid - Unified Requirements
# =================================================
# NOTE: torch and torchvision must be installed separately based on your CUDA version
# See README.md for installation instructions

# --- Core Deep Learning Framework ---
# torch==2.7.1+cpu/cu118/cu126/cu128  # Install separately based on CUDA version
# torchvision==0.22.1+cpu/cu118/cu126/cu128  # Install separately

# --- OpenMMLab Stack (Semantic Segmentation) ---
mmengine==0.10.4
openmim==0.3.9
mmcv==2.1.0  # Will be installed by MIM to match Torch/CUDA
mmsegmentation==1.2.2

# --- Data Handling & Processing ---
numpy==1.26.4
pandas==2.3.3
pillow==11.3.0
opencv-python==4.10.0.84

# --- ARTeFACT ---
datasets wandb transformers pytorch-lightning==1.9.2 albumentations torchmetrics torchinfo requests gradio

# --- Visualization ---
matplotlib==3.9.2
seaborn>=0.12.0

# --- Progress & Utilities ---
tqdm==4.66.4

# --- Configuration Management ---
PyYAML==6.0.2
hydra-core>=1.3.0
omegaconf>=2.3.0

# --- Model Backbones & Transformers ---
timm==0.9.16  # PyTorch Image Models (ConvNeXt, Swin, etc.)

# --- Text Processing (for data preprocessing) ---
ftfy==6.2.3
regex==2024.9.11

# --- Metrics & Evaluation ---
scikit-learn>=1.3.0
scipy==1.15.3

# --- Testing ---
pytest>=7.4.0
pytest-cov>=4.1.0

# --- Development Tools ---
black>=23.0.0  # Uncomment for code formatting
flake8>=6.0.0  # Uncomment for linting
pre-commit>=3.0.0  # Uncomment for Git hooks

# --- Optional: Experiment Tracking ---
# wandb>=0.15.0  # Uncomment for Weights & Biases integration
